+++ database/migrations/2024_12_04_021509_create_tbl_council_lecturers_table.php
@@ -13,13 +13,14 @@ return new class extends Migration
     //Bảng giảng viên trong hội đồng (liên kết giảng viên với hội đồng đánh giá).
     public function up()
     {
-        Schema::create('tbl_council_lecturers', function (Blueprint $table) {
-            $table->foreignId('council_id')->constrained('tbl_evaluation_councils')->onDelete('cascade');
-            $table->foreignId('lecturer_id')->constrained('tbl_lecturers')->onDelete('cascade');
-            $table->string('duty',100);
-            $table->enum('duty', ['Chủ tịch hội đồng', 'Ủy viên', 'Thư ký', 'Hỗ trợ kỹ thuật']);
-            $table->primary(['council_id', 'lecturer_id']);
-        });
+        if (!Schema::hasTable('tbl_council_lecturers')) {
+            Schema::create('tbl_council_lecturers', function (Blueprint $table) {
+                $table->foreignId('council_id')->constrained('tbl_evaluation_councils')->onDelete('cascade');
+                $table->foreignId('lecturer_id')->constrained('tbl_lecturers')->onDelete('cascade');
+                $table->enum('duty', ['Chủ tịch hội đồng', 'Ủy viên', 'Thư ký', 'Hỗ trợ kỹ thuật']);
+                $table->primary(['council_id', 'lecturer_id']);
+            });
+        }
     }
