+++ database/migrations/2024_12_04_021511_create_tbl_council_topics_table.php
@@ -13,11 +13,13 @@ return new class extends Migration
     //Bảng hội đồng - đề tài (liên kết các đề tài được đánh giá bởi hội đồng).
     public function up()
     {
-        Schema::create('tbl_council_topics', function (Blueprint $table) {
-            $table->foreignId('council_id')->constrained('tbl_evaluation_councils')->onDelete('cascade');
-            $table->foreignId('topic_id')->constrained('tbl_research_topics')->onDelete('cascade');
-            $table->primary(['council_id', 'topic_id']);
-        });
+        if (!Schema::hasTable('tbl_council_topics')) {
+            Schema::create('tbl_council_topics', function (Blueprint $table) {
+                $table->foreignId('council_id')->constrained('tbl_evaluation_councils')->onDelete('cascade');
+                $table->foreignId('topic_id')->constrained('tbl_research_topics')->onDelete('cascade');
+                $table->primary(['council_id', 'topic_id']);
+            });
+        }
     }
